refactor(contract): extract linearity axis into its own module (fix §14.1 budget)
The linearity additions pushed contract_decl.rs to 531 lines, over the 500-line
file-size budget (§14.1) that the Quality Gate enforces. Move the `Linearity`
enum, its `as_str`, and `linearity_from_word` into a new `contract_linearity.rs`
(42 lines); contract_decl.rs returns to 495 lines. No behavior change — the
15 contract_decl tests stay green and the file-size budget check now passes
(0 violations).
